Add ItemService with validation throughout and tag-juntion support

+++ MediaTracker/src/main/java/media/service/CategoryService.java
@@ -36,6 +36,9 @@ public class CategoryService {
 	}
 	
 	public void removeCategory(int categoryId)throws SQLException{
+		if(repo.findById(categoryId) == null) {
+			throw new IllegalArgumentException("Category with id " + categoryId + " does not exist.");
+		}
 		repo.delete(categoryId);
 	}
